Se realizo la funcion de Modificar y Eliminar tanto en el archivo como los registro que se tiene en memoria

+++ controllers/controller_image.py
@@ -28,18 +28,18 @@ class ImageController:
                 return True
         return False
     
-    def register_image(self, image_file, size, url):
+    def register_image(self, file_image, size, url):
 
-        image_name, extension = os.path.splitext(image_file)
+        image_name, extension = os.path.splitext(file_image)
         image_format = extension[1:].upper()
 
-        exsts_image = self.search_image(image_name.upper())
+        exists_image = self.search_image(image_name.upper())
 
         image_name = image_name.upper() + extension.lower()
         route_image = os.path.join(self.src_path, image_name)
         route_image2 = os.path.join(self.src_path2, image_name)
 
-        if os.path.exists(route_image) and exsts_image:
+        if os.path.exists(route_image) and exists_image:
             shutil.copy(route_image, self.dst_path)
             shutil.copy(route_image2, self.dst_path2)
             id_image = len(self.images) + 1
@@ -59,21 +59,23 @@ class ImageController:
             print(f"ID: {image.id}, Nombre: {image.file_name}, Size: {image.size} , Url: {image.url} ,\n Ruta1: {image.path},\n Ruta2: {image.path2}")
     
     
-    def modify_image(self, id_image, image_file):
+    def modify_image(self, id_image, file_image, size, url):
 
-        image_name, extension = os.path.splitext(image_file)
+        image_name, extension = os.path.splitext(file_image)
         image_name = image_name.upper() + extension.lower()
         new_image_path = os.path.join(self.dst_path, image_name)
         new_image_path2 = os.path.join(self.dst_path2, image_name)
 
         for image in self.images:
             if image.id == id_image:
-                current_image_path = os.path.join(image.path, image.name)
-                current_image_path2 = os.path.join(image.path2, image.name)
-                if os.path.exists(current_image_path):
+                current_image_path = os.path.join(image.path, image.file_name)
+                current_image_path2 = os.path.join(image.path2, image.file_name)
+                if os.path.exists(current_image_path) and os.path.exists(current_image_path2):
                     os.rename(current_image_path, new_image_path)
                     os.rename(current_image_path2, new_image_path2)
-                    image.name = image_name
+                    image.file_name = image_name
+                    image.size = size
+                    image.url = url
                     return True
                 else:
                     return False
@@ -83,8 +85,8 @@ class ImageController:
     def delete_image(self, id_image):
         for image in self.images:
             if image.id == id_image:
-                current_image_path = os.path.join(image.path, image.name)
-                current_image_path2 = os.path.join(image.path2, image.name)
+                current_image_path = os.path.join(image.path, image.file_name)
+                current_image_path2 = os.path.join(image.path2, image.file_name)
                 if os.path.exists(current_image_path):
                     os.remove(current_image_path)
                     os.remove(current_image_path2)
